Fix coc
SpaceVim, for some reason, only wants coc to use an older LSP that is
quite slow. Because we're loading coc already for autocompletion, we
don't need to use the lsp layer. Removing it fixes some of the issues.

Maps have to be moved to after command so they don't get overwritten by
spacevim